matched func_80037FE0_38BE0 1 attempt

+++ src/E770.c
@@ -36,7 +36,7 @@ typedef struct {
 } D_8008D348_Entry;
 
 extern D_8008D348_Entry D_8008D348_8DF48[];
-extern u8 func_80037FE0_38BE0(u8);
+extern s32 func_80037FE0_38BE0(u8);
 extern u8 func_80038000_38C00(u8);
 extern u8 func_80038058_38C58(u8);
 extern u8 func_80038070_38C70(void);
